+++ examples/optimizers.jl
@@ -1,6 +1,9 @@
-#=====================================#
+
 # MIP solvers (commercial, but fast)  #
-#=====================================#
+#=====================================##=====================================#
+
+# MIP solvers (commercial, but fast)  #
+
 function get_gurobi(; solver_log = true)
     GRB_ENV = Gurobi.Env()
     return optimizer_with_attributes(
@@ -20,11 +23,9 @@ function get_cplex(; solver_log = true)
         # "CPX_PARAM_MIPEMPHASIS" => 1, # Focus on optimality over feasibility 
         "CPX_PARAM_PREIND" => 1,
     )
-end
+end#======================================##======================================#
 
-#======================================#
 # MIP solvers (open-source, but slow)  #
-#======================================#
 
 function get_glpk(; solver_log = true)
     return JuMP.optimizer_with_attributes(GLPK.Optimizer, MOI.Silent() => !solver_log)
@@ -37,11 +38,9 @@ function get_highs()
         "presolve" => "on",
         "log_to_console" => true,
     )
-end
+end#========================================================##========================================================#
 
-#========================================================#
 # Continuous nonlinear programming solver (open-source)  #
-#========================================================#
 
 function get_ipopt(; solver_log = true)
     return JuMP.optimizer_with_attributes(
@@ -50,11 +49,10 @@ function get_ipopt(; solver_log = true)
         "sb" => "yes",
         "max_iter" => Int(1E4),
     )
-end
+end#=================================================================##=================================================================#
 
-#=================================================================#
 # Local mixed-integer nonlinear programming solver (open-source)  #
-#=================================================================#
+
 function get_juniper()
     return JuMP.optimizer_with_attributes(
         Juniper.Optimizer,